+++ Services/Ces/V2/Model/ListAlarmResponseAlarms.cs
@@ -53,11 +53,10 @@ namespace HuaweiCloud.SDK.Ces.V2.Model
         public List<ResourcesInListResp> Resources { get; set; }
 
         /// <summary>
-        /// 告警规则类型
+        /// 
         /// </summary>
         [JsonProperty("type", NullValueHandling = NullValueHandling.Ignore)]
-        public string Type { get; set; }
-
+        public AlarmType Type { get; set; }
         /// <summary>
         /// 告警开关
         /// </summary>